switch to uv for pip install
